Move some tests around to better match the current package layout.
